Prevent overrecycling on bad progs
This tweaks lfsr_mdir_commit_ to avoid overrecycling if we encounter a
bad prog (LFS_ERR_CORRUPT). This avoids compacting to the same block
twice, which risks an undetected prog error and breaks internal
invariants.

Note we still overrecycle if the relocation reason is a recycle
overflow.

---

This is an alternative solution to the previous overrecycling + shrub +
ckprog bug: Just make sure we don't compact to the same block twice!

After all, if we just got a bad prog, why are we trying to prog again?

(There are actually some arguments for multiple prog attempts, bus
errors for example, but I don't think that's a great excuse for littlefs
attempting multiple progs without user input.)

Even though this adds logic to lfsr_mdir_commit_, it ends up saving
code since we can drop the shrub discard pass:

           code          stack          ctx
  before: 37088           2304          636
  after:  37056 (-0.1%)   2304 (+0.0%)  636 (+0.0%)

Not that we _really_ care about this quantity of code. The real
motivation is 1. lowering the risk of a missed prog error, and
2. maintaining the never-compact-same-block invariant in case there
are other invariant-dependent bugs lurking around.