Generalized lfsr_file_crystallize_ for future write strategies
This function is actually pretty much the same in both the lazy and
eager crystallization write strategies. The main difference being the
nuances around the crystal_size parameter:

- lazy:  crystal_size => rough upper bound on crystal
- eager: crystal_size => strict lower bound on crystal

If we change these to an explicit crystal_min and crystal_max, we can
use lfsr_file_crystallize_ in both write strategies without changing the
logic.

It's out of scope right now, but this will help supporting both write
strategies in the future.

---

Unfortunately this added more code/stack that I was expecting:

           code          stack          ctx
  before: 36428           2248          636
  after:  36460 (+0.1%)   2280 (+1.4%)  636 (+0.0%)

I'm not exactly sure why, I guess the crystal_limit calculation is too
complex to const propagate the crystal_max=-1?

Maybe the LFS_NOINLINE is disabling certain cross-function
optimizations...