Reverted second resume-crystallization check
Gah! I'm not sure why I thought this code was so useless... Without it
we risk immediate recrystallization if the crystal heuristic pushes
crystal_start such that it overlaps the crystallizing block.

We may not make progress on our buffer, but triggering recrystallization
isn't great.

Considering this really doesn't add _that_ much code, I think this is a
case where we are better safe than sorry:

           code          stack          ctx
  before: 37092           2304          636
  after:  37140 (+0.1%)   2304 (+0.0%)  636 (+0.0%)

This logic only gets hit after we decide to allocate a new block, so
there's no risk of losing erased-state to potential fragments.

---

In benchmarking it also looks like this recoups most of the extra disk
usage introduced by lazy rewrite crystallization. So that's a good
thing... I think...