Reused block_start/end in crystallize math
Not sure how the duplication here went unnoticed.

Note we can _not_ reuse block_start/end for the buffer updates, since
those depend on the crystallized/aligned result.

No code changes though. The good news is the compiler is doing a good
job with the dense math in these functions.