Fixed buffer overflow when file caches are different sizes
This was a simple oversight, we weren't checking recipient file caches
when broadcasting sync!

Fixed by limiting the synced cache to the last n bytes that fit in the
recipient's cache. This is a bit more complicated than first n bytes,
but more intuitive/likely to be relevant to the recipient file.

Adds a bit of code/stack. In theory this shouldn't really affect the
stack, but lfsr_file_sync is a sensitive function on the stack hot-path:

           code          stack          ctx
  before: 37220           2288          636
  after:  37260 (+0.1%)   2296 (+0.3%)  636 (+0.0%)